Misc (re #1134): fixed compilation error in python module due to last change in session timer field from require_timer to use_timer. Thanks Beau Gunderson for the patch!
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@3352 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip-apps/src/python/_pjsua.c
@@ -2183,7 +2183,7 @@ static PyObject *py_pjsua_im_send(PyObject *pSelf, PyObject *pArgs)
     }
 
     status = pjsua_im_send(acc_id, &to, mime_type, &content, 
-			   &msg_data, (void *)user_data);
+			   &msg_data, (void*)(long)user_data);
     if (pool)
 	pj_pool_release(pool);
     
@@ -3710,7 +3710,7 @@ static PyObject *py_pjsua_call_send_im(PyObject *pSelf, PyObject *pArgs)
     }
     
     status = pjsua_call_send_im(call_id, mime_type, &content, 
-				&msg_data, (void *)user_data);
+				&msg_data, (void*)(long)user_data);
 
     if (pool)
         pj_pool_release(pool);
